orders and cart APIs created and verified (the earlier ones too)

+++ OnlineShoppingApplication/src/main/java/com/OnlineShoppingApp/Service/CartServiceImpl.java
@@ -1,6 +1,7 @@
 package com.OnlineShoppingApp.Service;
 
 import java.lang.StackWalker.Option;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,6 +14,7 @@ import com.OnlineShoppingApp.Entity.CartProduct;
 import com.OnlineShoppingApp.Entity.CurrentSession;
 import com.OnlineShoppingApp.Entity.Customer;
 import com.OnlineShoppingApp.Entity.Product;
+import com.OnlineShoppingApp.Enum.CartProductStatus;
 import com.OnlineShoppingApp.Exception.CartException;
 import com.OnlineShoppingApp.Exception.CustomerException;
 import com.OnlineShoppingApp.Exception.ProductException;
@@ -40,7 +42,7 @@ public class CartServiceImpl implements CartService{
 	private SessionDao sessionDao;
 	
 	@Override
-	public Cart addProductToCart(AddProductToCartDTO dto, String key) throws CartException,ProductException{
+	public CartProduct addProductToCart(AddProductToCartDTO dto, String key) throws CartException,ProductException{
 		CurrentSession validSession = sessionDao.findByUuid(key);
 		if(validSession!=null && validSession.getRole().toString().equals("CUSTOMER")) {
 			Optional<Cart> cartOpt = cartDao.findById(dto.getCartId());
@@ -55,10 +57,13 @@ public class CartServiceImpl implements CartService{
 			cartProduct.setProduct(prodOpt.get());
 			cartProduct.setQuantity(dto.getQuantity());
 			cartProduct.setCart(existingCart);
+			cartProduct.setCpStatus(CartProductStatus.UNORDERED);
 			
 			existingCart.getCartProductList().add(cartProduct);
 
-			return cartDao.save(existingCart);
+			Cart savedCart = cartDao.save(existingCart);
+			List<CartProduct> updatedCpList = savedCart.getCartProductList();
+			return updatedCpList.get(updatedCpList.size()-1);
 		}
 		throw new CartException("Invalid key or not of a customer.");
 	}
@@ -72,6 +77,9 @@ public class CartServiceImpl implements CartService{
 				throw new CartException("Invalid cart product id.");
 			}
 			CartProduct existingCp = cpOpt.get();
+			if(existingCp.getCpStatus().toString().equals(CartProductStatus.ORDERED.toString())) {
+				throw new CartException("This is an ordered cart product. This will be removed once your the order is delivered or canceled.");
+			}
 			cpDao.delete(existingCp);
 			return existingCp;
 		}
@@ -94,19 +102,23 @@ public class CartServiceImpl implements CartService{
 	}
 
 	@Override
-	public Cart removeAllProductsInCart(Integer customerId, String key) throws CartException,CustomerException{
+	public List<CartProduct> removeAllProductsInCart(Integer customerId, String key) throws CartException,CustomerException{
 		CurrentSession validSession = sessionDao.findByUuid(key);
 		if(validSession!=null && validSession.getRole().toString().equals("CUSTOMER")) {
 			Optional<Customer> customerOpt = customerDao.findById(customerId);
 			if(customerOpt.isPresent()) {
 				Customer existingCustomer = customerOpt.get();
 				List<CartProduct> cartProducts = existingCustomer.getCart().getCartProductList();
-				
+				List<CartProduct> deletedCartProducts = new ArrayList<>();
 				if(cartProducts.size()!=0) {
-					for(int i=0;i<cartProducts.size();i++) {
-						cpDao.delete(cartProducts.get(i));
+					while(cartProducts.size()>0) {
+						Optional<CartProduct> cpOpt = cpDao.findById(cartProducts.get(0).getCartProductId());
+						CartProduct existingCp = cpOpt.get();
+						deletedCartProducts.add(existingCp);
+						cartProducts.remove(0);
+						cpDao.delete(existingCp);
 					}
-					return cartDao.save(existingCustomer.getCart());
+				return deletedCartProducts;
 				}
 				throw new CartException("No product present in the cart of customer with id: "+customerId);
 			}
@@ -126,5 +138,23 @@ public class CartServiceImpl implements CartService{
 			throw new CartException("No product present in the cart of customer with id: "+customerId);
 		}
 		throw new CustomerException("Invalid customer id: "+customerId);
+	}
+
+	@Override
+	public Double findTotalCartPriceByCustomerId(Integer customerId) throws CartException {
+		Optional<Customer> customerOpt = customerDao.findById(customerId);
+		if(customerOpt.isPresent()) {
+			Customer existingCustomer = customerOpt.get();
+			List<CartProduct> customerCartProducts = existingCustomer.getCart().getCartProductList();
+			Double totalPrice = 0.0;
+			if(customerCartProducts.size()!=0) {
+				for(CartProduct cp:customerCartProducts) {
+					totalPrice=totalPrice + (cp.getQuantity()*cp.getProduct().getPrice());
+				}
+				return totalPrice;
+			}
+			return totalPrice;
+		}
+		throw new CartException("Invalid customer id: "+customerId);
 	}
 }
